third_party/blink/web_tests/fast/dom/shadow/shadowhost-keyframes-expected.txt

CONSOLE WARNING: line 30: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test if @keyframes works for shadow host when using :host. i.e. crbug.com/332577

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getColor(document.getElementById('host')) is "rgb(0, 128, 0)"
